+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Add location permission set. Fixes UIORG-76.
 * Better parens handling for institutions. Refs UIORG-69.
 * Include location-count on libraries page. Refs UIORG-66.
+* Refresh lookup tables on mount. Refs UIORG-69.
 
 ## [2.2.0](https://github.com/folio-org/ui-organization/tree/v2.2.0) (2017-09-01)
 [Full Changelog](https://github.com/folio-org/ui-organization/compare/v2.1.0...v2.2.0)
@@ -90,4 +91,3 @@
 ## [0.1.0](https://github.com/folio-org/ui-organization/tree/v0.1.0) (2017-05-05)
 
 * The first formal release.
-